requirements.txt
streamlit>=1.35
langchain>=0.3
langchain-core>=0.3
langchain-groq>=0.2
langchain-chroma>=0.1
langgraph>=0.2
chromadb>=0.5
python-dotenv>=1.0
typing-extensions>=4.9
sentence-transformers>=3.0.1
langchain-huggingface>=0.0.3